Move Transformations from output sidebar to input sidebar transforms tab
